requirements.txt
openai~=2.7.2
python-dotenv~=1.1.1
faiss-cpu~=1.12.0
numpy~=1.26.4
pandas~=2.2.3
sentence-transformers~=5.1.2
natasha~=1.6.0
datasets~=2.19.1
transformers~=4.45.2
yaml~=0.2.5
pyyaml~=6.0.2
scikit-learn~=1.6.1